requirements.txt
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
RapidFuzz==3.14.3
rich==14.2.0
six==1.17.0
